# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Electricity Waste Detection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DeviceType = (typeof DeviceType)[keyof typeof DeviceType];

export const DeviceType = {
  hvac: "hvac",
  lighting: "lighting",
  appliance: "appliance",
  ev_charger: "ev_charger",
  industrial: "industrial",
  other: "other",
} as const;

export type DeviceStatus = (typeof DeviceStatus)[keyof typeof DeviceStatus];

export const DeviceStatus = {
  online: "online",
  offline: "offline",
  warning: "warning",
} as const;

export interface Device {
  id: number;
  name: string;
  location: string;
  type: DeviceType;
  status: DeviceStatus;
  /** Rated power in watts */
  powerRatingW: number;
  /** Current consumption in watts */
  currentWatts: number;
  createdAt: string;
}

export type CreateDeviceBodyType =
  (typeof CreateDeviceBodyType)[keyof typeof CreateDeviceBodyType];

export const CreateDeviceBodyType = {
  hvac: "hvac",
  lighting: "lighting",
  appliance: "appliance",
  ev_charger: "ev_charger",
  industrial: "industrial",
  other: "other",
} as const;

export interface CreateDeviceBody {
  name: string;
  location: string;
  type: CreateDeviceBodyType;
  powerRatingW: number;
}

export type UpdateDeviceBodyType =
  (typeof UpdateDeviceBodyType)[keyof typeof UpdateDeviceBodyType];

export const UpdateDeviceBodyType = {
  hvac: "hvac",
  lighting: "lighting",
  appliance: "appliance",
  ev_charger: "ev_charger",
  industrial: "industrial",
  other: "other",
} as const;

export type UpdateDeviceBodyStatus =
  (typeof UpdateDeviceBodyStatus)[keyof typeof UpdateDeviceBodyStatus];

export const UpdateDeviceBodyStatus = {
  online: "online",
  offline: "offline",
  warning: "warning",
} as const;

export interface UpdateDeviceBody {
  name?: string;
  location?: string;
  type?: UpdateDeviceBodyType;
  status?: UpdateDeviceBodyStatus;
  powerRatingW?: number;
}

export interface Reading {
  id: number;
  deviceId: number;
  deviceName: string;
  wattsConsumed: number;
  kwhConsumed: number;
  isWasteful: boolean;
  recordedAt: string;
}

export interface CreateReadingBody {
  deviceId: number;
  wattsConsumed: number;
}

export type AlertType = (typeof AlertType)[keyof typeof AlertType];

export const AlertType = {
  overconsumption: "overconsumption",
  idle_waste: "idle_waste",
  off_hours_usage: "off_hours_usage",
  spike_detected: "spike_detected",
  threshold_exceeded: "threshold_exceeded",
} as const;

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type AlertStatus = (typeof AlertStatus)[keyof typeof AlertStatus];

export const AlertStatus = {
  active: "active",
  resolved: "resolved",
  dismissed: "dismissed",
} as const;

export interface Alert {
  id: number;
  deviceId: number;
  deviceName: string;
  type: AlertType;
  severity: AlertSeverity;
  message: string;
  status: AlertStatus;
  estimatedWasteKwh: number;
  estimatedCostUsd: number;
  createdAt: string;
  resolvedAt?: string | null;
}

export interface DashboardSummary {
  totalDevices: number;
  onlineDevices: number;
  activeAlerts: number;
  criticalAlerts: number;
  totalConsumptionKwhToday: number;
  wasteKwhToday: number;
  estimatedDailyCostUsd: number;
  estimatedMonthlySavingsUsd: number;
  wastePercentage: number;
}

export interface ConsumptionDataPoint {
  label: string;
  totalKwh: number;
  wasteKwh: number;
  costUsd: number;
}

export interface DeviceWaste {
  deviceId: number;
  deviceName: string;
  location: string;
  wasteKwh: number;
  wastePercentage: number;
  estimatedWasteCostUsd: number;
}

export interface SavingsSummary {
  totalSavedKwh: number;
  totalSavedUsd: number;
  resolvedAlertsCount: number;
  thisMonthSavedUsd: number;
  lastMonthSavedUsd: number;
  /** Percentage change vs last month */
  savingsTrend: number;
}

export type ListReadingsParams = {
  deviceId?: number;
  limit?: number;
};

export type ListAlertsParams = {
  status?: ListAlertsStatus;
};

export type ListAlertsStatus =
  (typeof ListAlertsStatus)[keyof typeof ListAlertsStatus];

export const ListAlertsStatus = {
  active: "active",
  resolved: "resolved",
  dismissed: "dismissed",
} as const;

export type GetConsumptionChartParams = {
  period?: GetConsumptionChartPeriod;
};

export type GetConsumptionChartPeriod =
  (typeof GetConsumptionChartPeriod)[keyof typeof GetConsumptionChartPeriod];

export const GetConsumptionChartPeriod = {
  day: "day",
  week: "week",
  month: "month",
} as const;