# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Electricity Waste Detection API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all monitored devices
 */
export const ListDevicesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  type: zod.enum([
    "hvac",
    "lighting",
    "appliance",
    "ev_charger",
    "industrial",
    "other",
  ]),
  status: zod.enum(["online", "offline", "warning"]),
  powerRatingW: zod.number().describe("Rated power in watts"),
  currentWatts: zod.number().describe("Current consumption in watts"),
  createdAt: zod.coerce.date(),
});
export const ListDevicesResponse = zod.array(ListDevicesResponseItem);

/**
 * @summary Register a new device
 */
export const CreateDeviceBody = zod.object({
  name: zod.string(),
  location: zod.string(),
  type: zod.enum([
    "hvac",
    "lighting",
    "appliance",
    "ev_charger",
    "industrial",
    "other",
  ]),
  powerRatingW: zod.number(),
});

/**
 * @summary Get a specific device
 */
export const GetDeviceParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDeviceResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  type: zod.enum([
    "hvac",
    "lighting",
    "appliance",
    "ev_charger",
    "industrial",
    "other",
  ]),
  status: zod.enum(["online", "offline", "warning"]),
  powerRatingW: zod.number().describe("Rated power in watts"),
  currentWatts: zod.number().describe("Current consumption in watts"),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update a device
 */
export const UpdateDeviceParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateDeviceBody = zod.object({
  name: zod.string().optional(),
  location: zod.string().optional(),
  type: zod
    .enum([
      "hvac",
      "lighting",
      "appliance",
      "ev_charger",
      "industrial",
      "other",
    ])
    .optional(),
  status: zod.enum(["online", "offline", "warning"]).optional(),
  powerRatingW: zod.number().optional(),
});

export const UpdateDeviceResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  type: zod.enum([
    "hvac",
    "lighting",
    "appliance",
    "ev_charger",
    "industrial",
    "other",
  ]),
  status: zod.enum(["online", "offline", "warning"]),
  powerRatingW: zod.number().describe("Rated power in watts"),
  currentWatts: zod.number().describe("Current consumption in watts"),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Remove a device
 */
export const DeleteDeviceParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List energy readings
 */
export const listReadingsQueryLimitDefault = 50;

export const ListReadingsQueryParams = zod.object({
  deviceId: zod.coerce.number().optional(),
  limit: zod.coerce.number().default(listReadingsQueryLimitDefault),
});

export const ListReadingsResponseItem = zod.object({
  id: zod.number(),
  deviceId: zod.number(),
  deviceName: zod.string(),
  wattsConsumed: zod.number(),
  kwhConsumed: zod.number(),
  isWasteful: zod.boolean(),
  recordedAt: zod.coerce.date(),
});
export const ListReadingsResponse = zod.array(ListReadingsResponseItem);

/**
 * @summary Record a new energy reading
 */
export const CreateReadingBody = zod.object({
  deviceId: zod.number(),
  wattsConsumed: zod.number(),
});

/**
 * @summary List waste alerts
 */
export const ListAlertsQueryParams = zod.object({
  status: zod.enum(["active", "resolved", "dismissed"]).optional(),
});

export const ListAlertsResponseItem = zod.object({
  id: zod.number(),
  deviceId: zod.number(),
  deviceName: zod.string(),
  type: zod.enum([
    "overconsumption",
    "idle_waste",
    "off_hours_usage",
    "spike_detected",
    "threshold_exceeded",
  ]),
  severity: zod.enum(["low", "medium", "high", "critical"]),
  message: zod.string(),
  status: zod.enum(["active", "resolved", "dismissed"]),
  estimatedWasteKwh: zod.number(),
  estimatedCostUsd: zod.number(),
  createdAt: zod.coerce.date(),
  resolvedAt: zod.coerce.date().nullish(),
});
export const ListAlertsResponse = zod.array(ListAlertsResponseItem);

/**
 * @summary Resolve an alert
 */
export const ResolveAlertParams = zod.object({
  id: zod.coerce.number(),
});

export const ResolveAlertResponse = zod.object({
  id: zod.number(),
  deviceId: zod.number(),
  deviceName: zod.string(),
  type: zod.enum([
    "overconsumption",
    "idle_waste",
    "off_hours_usage",
    "spike_detected",
    "threshold_exceeded",
  ]),
  severity: zod.enum(["low", "medium", "high", "critical"]),
  message: zod.string(),
  status: zod.enum(["active", "resolved", "dismissed"]),
  estimatedWasteKwh: zod.number(),
  estimatedCostUsd: zod.number(),
  createdAt: zod.coerce.date(),
  resolvedAt: zod.coerce.date().nullish(),
});

/**
 * @summary Dismiss an alert
 */
export const DismissAlertParams = zod.object({
  id: zod.coerce.number(),
});

export const DismissAlertResponse = zod.object({
  id: zod.number(),
  deviceId: zod.number(),
  deviceName: zod.string(),
  type: zod.enum([
    "overconsumption",
    "idle_waste",
    "off_hours_usage",
    "spike_detected",
    "threshold_exceeded",
  ]),
  severity: zod.enum(["low", "medium", "high", "critical"]),
  message: zod.string(),
  status: zod.enum(["active", "resolved", "dismissed"]),
  estimatedWasteKwh: zod.number(),
  estimatedCostUsd: zod.number(),
  createdAt: zod.coerce.date(),
  resolvedAt: zod.coerce.date().nullish(),
});

/**
 * @summary Get dashboard overview stats
 */
export const GetDashboardSummaryResponse = zod.object({
  totalDevices: zod.number(),
  onlineDevices: zod.number(),
  activeAlerts: zod.number(),
  criticalAlerts: zod.number(),
  totalConsumptionKwhToday: zod.number(),
  wasteKwhToday: zod.number(),
  estimatedDailyCostUsd: zod.number(),
  estimatedMonthlySavingsUsd: zod.number(),
  wastePercentage: zod.number(),
});

/**
 * @summary Get hourly/daily consumption data for chart
 */
export const getConsumptionChartQueryPeriodDefault = `week`;

export const GetConsumptionChartQueryParams = zod.object({
  period: zod
    .enum(["day", "week", "month"])
    .default(getConsumptionChartQueryPeriodDefault),
});

export const GetConsumptionChartResponseItem = zod.object({
  label: zod.string(),
  totalKwh: zod.number(),
  wasteKwh: zod.number(),
  costUsd: zod.number(),
});
export const GetConsumptionChartResponse = zod.array(
  GetConsumptionChartResponseItem,
);

/**
 * @summary Get waste breakdown by device
 */
export const GetWasteByDeviceResponseItem = zod.object({
  deviceId: zod.number(),
  deviceName: zod.string(),
  location: zod.string(),
  wasteKwh: zod.number(),
  wastePercentage: zod.number(),
  estimatedWasteCostUsd: zod.number(),
});
export const GetWasteByDeviceResponse = zod.array(GetWasteByDeviceResponseItem);

/**
 * @summary Get estimated cost savings from resolved waste
 */
export const GetSavingsSummaryResponse = zod.object({
  totalSavedKwh: zod.number(),
  totalSavedUsd: zod.number(),
  resolvedAlertsCount: zod.number(),
  thisMonthSavedUsd: zod.number(),
  lastMonthSavedUsd: zod.number(),
  savingsTrend: zod.number().describe("Percentage change vs last month"),
});