# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Electricity Waste Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from "./alertSeverity";
import type { AlertStatus } from "./alertStatus";
import type { AlertType } from "./alertType";

export interface Alert {
  id: number;
  deviceId: number;
  deviceName: string;
  type: AlertType;
  severity: AlertSeverity;
  message: string;
  status: AlertStatus;
  estimatedWasteKwh: number;
  estimatedCostUsd: number;
  createdAt: Date;
  resolvedAt?: Date | null;
}