# lib/api-zod/src/generated/types/createReadingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Electricity Waste Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReadingBody {
  deviceId: number;
  wattsConsumed: number;
}