# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Electricity Waste Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalDevices: number;
  onlineDevices: number;
  activeAlerts: number;
  criticalAlerts: number;
  totalConsumptionKwhToday: number;
  wasteKwhToday: number;
  estimatedDailyCostUsd: number;
  estimatedMonthlySavingsUsd: number;
  wastePercentage: number;
}