# lib/api-zod/src/generated/types/reading.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Electricity Waste Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface Reading {
  id: number;
  deviceId: number;
  deviceName: string;
  wattsConsumed: number;
  kwhConsumed: number;
  isWasteful: boolean;
  recordedAt: Date;
}